03- Mastering OpenAI APIs/06- Embeddings/requirements.txt
openai
python-dotenv
fastapi==0.95.1
uvicorn==0.20.0
pytube
python-multipart
openai[datalib]
matplotlib
plotly
scipy
numpy
pandas
scikit-learn
tiktoken